Restructuring and SEO bug fixes
