+++ src/components/seo/config.ts
@@ -1,5 +1,6 @@
 export default {
   title: "Mara - Belarusians NL",
   description: "Official website of Belarusian diaspora in the Netherlands - MARA",
-  siteUrl: "https://www.belarusians.nl",
+  siteUrl: "https://belarusians.nl",
+  imagePath: "/logo.jpeg",
 };
