--- src/pages/_app.js
@@ -1,4 +1,0 @@
-import { appWithTranslation } from 'next-i18next';
-import IndexPage from "./index";
-
-export default appWithTranslation(IndexPage);
