+++ src/pages/_app.tsx
@@ -1,0 +1,13 @@
+import { AppProps } from "next/app";
+import { appWithTranslation } from "next-i18next";
+import { themeClass } from "../components/sprinkles.css";
+
+function MyApp({ Component, pageProps }: AppProps) {
+  return (
+    <div className={themeClass}>
+      <Component {...pageProps} />
+    </div>
+  );
+}
+
+export default appWithTranslation(MyApp);
